PartyState & BattleScene Awake SetAlly
PartyState 필드 제작 및 Setter
기반으로 BattleScene에서 RuntimeState 의 PartyState와 소통하여 현재 체력으로 시작할 수 있도록 변경
탈리스만 충전 횟수 & 탐험 지속 버프는 스텁으로 유지

+++ Assets/Scripts/BattleScene/BattleUnit/AllyUnitUIcontroller.cs
@@ -13,7 +13,7 @@ public class AllyUnitUIcontroller : BattleUnitUIcontroller
     [SerializeField] private Text spd;
 
     [Header("Health")]
-    [SerializeField] private float maxHealth;
+    [SerializeField] private int maxHealth;
 
     public override void SetUpUI(BattleUnitBase unit )
     {
@@ -26,14 +26,14 @@ public class AllyUnitUIcontroller : BattleUnitUIcontroller
         base.SetUpUI(unit);
     }
 
-    public override void SetHealth(float currentHealth)
+    public override void SetHealth(int currentHealth)
     {
         float sliderValue = currentHealth / maxHealth;
         healthSlider.value = sliderValue;
         healthTxt.text = $"{currentHealth} / {maxHealth}";
     }
 
-    public void SetStatTexts(float attack, float shield, float speed)
+    public void SetStatTexts(int attack, int shield, int speed)
     {
         atk.text = attack.ToString();
         shd.text = shield.ToString();
